backup

+++ OptionsWebSite/Controllers/YearTermsController.cs
@@ -33,6 +33,16 @@ namespace OptionsWebSite.Controllers
             {
                 return HttpNotFound();
             }
+
+            var e = db.YearTerms.Find(id);
+       
+            ViewBag.q = "";
+
+            if (e.Term == 10) ViewBag.q += "Winter";
+            else if (e.Term == 20) ViewBag.q += "Spring/Summer";
+            else if (e.Term == 30) ViewBag.q += "Fall";
+            else ViewBag.q += ":(";
+            
             return View(yearTerm);
         }
 
@@ -71,6 +81,20 @@ namespace OptionsWebSite.Controllers
             {
                 return HttpNotFound();
             }
+
+            IList<YearTerm> yearTerms = db.YearTerms.ToList();
+            var name = new Dictionary<int, string>();
+
+            name[10] = "Winter";
+            name[20] = "Spring/Summer";
+            name[30] = "Fall";
+            IEnumerable<SelectListItem> selectList = from year in yearTerms
+                                                     select new SelectListItem
+                                                     {
+                                                         Text = name[year.Term],
+                                                         Value = year.Term.ToString()
+                                                     };
+            ViewBag.Term = selectList;
             return View(yearTerm);
         }
 
@@ -81,14 +105,42 @@ namespace OptionsWebSite.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "YearTermId,Year,Term,IsDefault")] YearTerm yearTerm)
         {
-            if (ModelState.IsValid)
+            if (yearTerm.Term != 10 && yearTerm.Term != 20 && yearTerm.Term != 30)
             {
-                db.Entry(yearTerm).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                ViewBag.Error = "Invalid value for term";
+                return View(yearTerm);
             }
-            return View(yearTerm);
-        }
+            var q = from y in db.YearTerms
+                        where y.IsDefault == true
+                        select y;
+                var s = q.ToArray();
+                if (yearTerm.IsDefault == false && s.Length <= 1 && yearTerm.YearTermId == s[0].YearTermId)
+                {
+                    ModelState.AddModelError("IsDefault", "There is no current default");
+                        
+                     }
+                if (yearTerm.IsDefault)
+                {
+                
+                    foreach (var year in db.YearTerms.Where(a=>a.IsDefault))
+                    {
+                        if (year.YearTermId != yearTerm.YearTermId)
+                            year.IsDefault = false;
+                    }
+                }
+
+            if (ModelState.IsValid)
+                {
+                
+                    YearTerm yt = db.YearTerms.Find(yearTerm.YearTermId);
+                    yt.Year = yearTerm.Year;
+                    yt.Term = yearTerm.Term;
+                    yt.IsDefault = yearTerm.IsDefault;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+                return View(yearTerm);
+            }
 
         // GET: YearTerms/Delete/5
         public ActionResult Delete(int? id)
@@ -102,6 +154,16 @@ namespace OptionsWebSite.Controllers
             {
                 return HttpNotFound();
             }
+
+            var e = db.YearTerms.Find(id);
+
+            ViewBag.q = "";
+
+            if (e.Term == 10) ViewBag.q += "Winter";
+            else if (e.Term == 20) ViewBag.q += "Spring/Summer";
+            else if (e.Term == 30) ViewBag.q += "Fall";
+            else ViewBag.q += ":(";
+
             return View(yearTerm);
         }
 
@@ -111,20 +173,22 @@ namespace OptionsWebSite.Controllers
         public ActionResult DeleteConfirmed(int id)
         {
             var choices = db.Choices.Where(a => a.YearTermId == id).Select(a => a.ChoiceId);
-                
-           foreach (var a in choices)
-            {
-                Choice c = db.Choices.Find(a);
-                db.Choices.Remove(c);
-            }
-            
-
             YearTerm yearTerm = db.YearTerms.Find(id);
-            db.YearTerms.Remove(yearTerm);
-            db.SaveChanges();
 
-            
-            return RedirectToAction("Index");
+            if ( choices.Count() != 0 )
+            {
+                ViewBag.error = "A choice with this year already exists";
+            }
+            else
+            {
+                db.YearTerms.Remove(yearTerm);
+                db.SaveChanges();
+                return RedirectToAction("Index");
+            }
+
+
+            return View(yearTerm);
+
         }
 
         protected override void Dispose(bool disposing)
